+++ packages/xstate-store-react/src/index.ts
@@ -1,17 +1,19 @@
+export * from '@xstate/store';
+
 import { useCallback, useRef, useSyncExternalStore } from 'react';
 import {
-  AnyStore,
-  StoreContext,
-  EventPayloadMap,
-  StoreConfig,
-  Store,
-  ExtractEvents,
-  Readable,
-  AnyAtom,
-  BaseAtom,
-  StoreSnapshot
-} from './types';
-import { createStore } from './store';
+  type AnyStore,
+  type StoreContext,
+  type EventPayloadMap,
+  type StoreConfig,
+  type Store,
+  type ExtractEvents,
+  type Readable,
+  type AnyAtom,
+  type BaseAtom,
+  type StoreSnapshot,
+  createStore
+} from '@xstate/store';
 
 function defaultCompare<T>(a: T | undefined, b: T) {
   return a === b;
@@ -39,7 +41,6 @@ function useSelectorWithCompare<TStore extends Readable<any>, T>(
  * A React hook that subscribes to the `store` and selects a value from the
  * store's snapshot via a selector function, with an optional compare function.
  *
- * @deprecated Use `useSelector` from `@xstate/store-react` instead.
  * @example
  *
  * ```ts
@@ -111,7 +112,6 @@ export function useSelector<TStore extends Readable<any>, T>(
   );
 }
 
-/** @deprecated Use `useStore` from `@xstate/store-react` instead. */
 export const useStore: {
   <
     TContext extends StoreContext,
@@ -145,7 +145,6 @@ export const useStore: {
  * A React hook that subscribes to the `atom` and returns the current value of
  * the atom.
  *
- * @deprecated Use `useAtom` from `@xstate/store-react` instead.
  * @example
  *
  * ```ts
@@ -189,7 +188,6 @@ export function useAtom(
  * Creates a custom hook that returns the selected value and the store from a
  * store configuration object.
  *
- * @deprecated Use `createStoreHook` from `@xstate/store-react` instead.
  * @example
  *
  * ```ts